requirements.txt
celery[redis]
click
dateparser
django>=4.0
django-celery-beat
django-environ
django-filter
django-stubs
djangorestframework
djangorestframework-stubs
gunicorn
mypy
psycopg2-binary
pylint
pylint-django
pyTelegramBotAPI
requests
types-dateparser
whitenoise
